proteccion de rutas check
