+++ src/app/admin/dashboard/page.tsx
@@ -1,0 +1,10 @@
+import AdminDashboardView from '@/views/AdminView/AdminView'
+import React from 'react'
+
+function AdminDashboardPage() {
+  return (
+    <AdminDashboardView />
+  )
+}
+
+export default AdminDashboardPage
