--- src/app/admin/page.tsx
@@ -1,10 +1,0 @@
-import React from 'react'
-import DashboardHeader from '@/components/specific/admin/AdminHeader/AdminHeader'
-
-const page = () => {
-  return (
-    <DashboardHeader/>
-  )
-}
-
-export default page
