+++ src/app/employee/dashboard/page.tsx
@@ -1,0 +1,10 @@
+import EmployeeDashboardView from '@/views/EmployeeView/EmployeeView'
+import React from 'react'
+
+function EmployeeDashboardPage() {
+  return (
+    <EmployeeDashboardView />
+  )
+}
+
+export default EmployeeDashboardPage
