--- src/app/employee/inicio/page.tsx
@@ -1,7 +1,0 @@
-import EmployeeDashboard from "@/components/specific/Dashboard/employeeDashboard";
-
-const Page = () => {
-  return <EmployeeDashboard />;
-};
-
-export default Page;
