+++ src/components/specific/Admin/AdminDashboard/AdminDashboard.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import AdminHeader from '@/components/specific/Admin/AdminHeader/AdminHeader'
+
+function AdminDashboard() {
+  return (
+    <AdminHeader />
+  )
+}
+
+export default AdminDashboard
