+++ src/components/specific/Employee/EmployeeHeader/EmployeeHeader.tsx
@@ -14,7 +14,7 @@ export default function EmployeeHeader() {
 
       {/* Navegación */}
       <nav className="space-y-3">
-        <NavItem href="/employee/inicio" icon={Home} label="Inicio" />
+        <NavItem href="/employee/dashboard" icon={Home} label="Inicio" />
         <NavItem href="/employee/reservations" icon={Clipboard} label="Reservas" />
         <NavItem href="/employee/orders" icon={ClipboardList} label="Órdenes" />
         <NavItem href="/employee/users" icon={Users} label="Usuarios" />
