+++ src/components/specific/Login/Login.tsx
@@ -1,49 +1,60 @@
 "use client";
- 
+
 import { useState } from "react";
 import { useUserContext } from "@/context/UserContext";
 import { login } from "@/helpers/auth.helper";
 import { useRouter } from "next/navigation";
 import Cookies from "js-cookie";
- 
+
 const Login = () => {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
   const [error, setError] = useState("");
   const { setUser } = useUserContext();
   const router = useRouter();
- 
+
   const handleLogin = async (e: React.FormEvent) => {
     e.preventDefault();
+    console.log("Login attempt with:", { email, password });
+
     try {
       const response = await login({ email, password });
- 
-      if (response) {        
- 
+
+      console.log("Response from server:", response.data);
+
+      if (response && response.data.token && response.data.user) {
         const token = response.data.token;
-        console.log(token);
- 
         const user = response.data.user;
- 
+
         Cookies.set("token", token, { expires: 7 });
+        console.log("Token stored in cookies:", Cookies.get("token"));
+
         setUser(user);
         localStorage.setItem("user", JSON.stringify(user));
+        console.log("User stored in localStorage:", localStorage.getItem("user"));
+
+        console.log("User Role:", user.role);
+
         if (user.role === "worker") {
-          router.push("/employee/inicio"); // Redirige al dashboard del empleado
+          console.log("Redirecting to /employee/dashboard");
+          router.push("/employee/dashboard");
+        } else if (user.role === "admin") {
+          console.log("Redirecting to /admin/dashboard");
+          router.push("/admin/dashboard");
         } else {
-          router.push("/profile"); // Redirige al perfil si no es Worker
+          console.log("Redirecting to /profile");
+          router.push("/profile");
         }
-      }
-       else {
+      } else {
+        console.log("Login failed: Invalid credentials");
         setError("Credenciales incorrectas");
       }
     } catch (error) {
       console.error("Login Error:", error);
       setError("Error al iniciar sesión");
-    }  
- 
+    }
   };
- 
+
   return (
     <form onSubmit={handleLogin} className="p-6 space-y-6 max-w-sm mx-auto">
       <div>
@@ -60,11 +71,9 @@ const Login = () => {
           className="input-field"
           required
         />
-        {error && (
-          <span className="text-red-500 text-sm mt-1 block">{error}</span>
-        )}
+        {error && <span className="text-red-500 text-sm mt-1 block">{error}</span>}
       </div>
- 
+
       <div>
         <label htmlFor="password" className="text-body mb-1 block">
           Contraseña
@@ -80,11 +89,11 @@ const Login = () => {
           required
         />
       </div>
- 
-      <button type="submit" className="button-primary" onClick={handleLogin}>
+
+      <button type="submit" className="button-primary">
         Iniciar sesión
       </button>
- 
+
       <button
         type="button"
         onClick={() => (window.location.href = "/api/auth/login")}
@@ -95,5 +104,5 @@ const Login = () => {
     </form>
   );
 };
- 
+
 export default Login;
