+++ src/views/AdminView/AdminView.tsx
@@ -1,0 +1,10 @@
+import AdminDashboard from '@/components/specific/Admin/AdminDashboard/AdminDashboard'
+import React from 'react'
+
+const AdminDashboardView = () => {
+  return (
+    <AdminDashboard />
+  )
+}
+
+export default AdminDashboardView
