--- src/views/AdminView/inicio/AdminView.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-import Card from '@/components/admin/cards/cards'
-const AdminView = () => {
-  return (
-    <Card/>
-  )
-}
-
-export default AdminView
