+++ src/views/EmployeeView/EmployeeView.tsx
@@ -1,0 +1,11 @@
+
+import EmployeeDashboard from '@/components/specific/Employee/EmployeeDashboard/EmployeeDashboard'
+import React from 'react'
+
+const EmployeeDashboardView = () => {
+  return (
+    <EmployeeDashboard />
+  )
+}
+
+export default EmployeeDashboardView
